requirements.txt
openai==1.75.0
python-dotenv
ipykernel
httpx